@@ -8,12 +8,13 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "gh-pages": "^6.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-fast-marquee": "^1.6.5",
     "react-helmet": "^6.1.0",
     "react-scripts": "4.0.3",
-    "react-scroll": "^1.8.4",
-    "react-slick": "^0.28.1",
+    "react-scroll": "^1.9.0",
+    "react-slick": "^0.30.2",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.1",
     "web-vitals": "^1.1.2"
