@@ -1,5 +1,4 @@
 import React from "react";
-import Slider from "react-slick";
 import styled from "styled-components";
 // Assets
 import ClientLogo01 from "../../assets/img/clients/logo01.svg";
@@ -8,73 +7,42 @@ import ClientLogo03 from "../../assets/img/clients/logo03.svg";
 import ClientLogo04 from "../../assets/img/clients/logo04.svg";
 import ClientLogo05 from "../../assets/img/clients/logo05.svg";
 import ClientLogo06 from "../../assets/img/clients/logo06.svg";
+import Marquee from "react-fast-marquee";
 
 export default function ClientSlider() {
-  const settings = {
-    infinite: true,
-    speed: 500,
-    slidesToShow: 6,
-    slidesToScroll: 2,
-    arrows: false,
-    responsive: [
-      {
-        breakpoint: 1024,
-        settings: {
-          slidesToShow: 4,
-          slidesToScroll: 2,
-        },
-      },
-      {
-        breakpoint: 600,
-        settings: {
-          slidesToShow: 3,
-          slidesToScroll: 3,
-        },
-      },
-      {
-        breakpoint: 480,
-        settings: {
-          slidesToShow: 2,
-          slidesToScroll: 2,
-        },
-      },
-    ],
-  };
   return (
-    <div>
-      <Slider {...settings}>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo01} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo02} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo03} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo04} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo05} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo06} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo03} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo04} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo01} alt="client logo" />
-        </LogoWrapper>
-        <LogoWrapper className="flexCenter">
-          <ImgStyle src={ClientLogo02} alt="client logo" />
-        </LogoWrapper>
-      </Slider>
-    </div>
+    <Marquee pauseOnHover={true} pauseOnClick={true}>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo01} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo02} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo03} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo04} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo05} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo06} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo03} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo04} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo01} alt="client logo" />
+      </LogoWrapper>
+      <LogoWrapper className="flexCenter">
+        <ImgStyle src={ClientLogo02} alt="client logo" />
+      </LogoWrapper>
+    </Marquee>
   );
 }
 
