@@ -6,7 +6,7 @@ import MonitorIcon from "../../assets/svg/Services/MonitorIcon";
 import BrowserIcon from "../../assets/svg/Services/BrowserIcon";
 import PrinterIcon from "../../assets/svg/Services/PrinterIcon";
 
-export default function ServiceBox({icon, title, subtitle}) {
+export default function ServiceBox({ icon, title, subtitle }) {
   let getIcon;
 
   switch (icon) {
@@ -27,7 +27,6 @@ export default function ServiceBox({icon, title, subtitle}) {
       break;
   }
 
-
   return (
     <Wrapper className="flex flexColumn">
       <IconStyle>{getIcon}</IconStyle>
@@ -39,6 +38,9 @@ export default function ServiceBox({icon, title, subtitle}) {
 
 const Wrapper = styled.div`
   width: 100%;
+  @media (max-width: 860px) {
+    align-items: center;
+  }
 `;
 const IconStyle = styled.div`
   @media (max-width: 860px) {
@@ -48,7 +50,6 @@ const IconStyle = styled.div`
 const TitleStyle = styled.h2`
   width: 100%;
   max-width: 300px;
-  margin: 0 auto;
   padding: 40px 0;
   @media (max-width: 860px) {
     padding: 20px 0;
@@ -57,5 +58,4 @@ const TitleStyle = styled.h2`
 const SubtitleStyle = styled.p`
   width: 100%;
   max-width: 300px;
-  margin: 0 auto;
 `;
